Issue #no-48 onDraw has been change with globalize class
